+++ CHANGELOG.md
@@ -10,17 +10,18 @@
 * New: Verify validity of admin color scheme on save and on retrieval
     * New: Add `is_valid_admin_color_scheme()` to check if a given admin color scheme is valid
     * Change: Modify `get_forced_color_scheme()` to not return an invalid admin color scheme
-    * Change: Modify `set_forced_admin_color()` to not save an invalid admin color scheme
+    * Change: Modify `set_forced_color_scheme()` to not save an invalid admin color scheme
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add to it)
 * Change: Output currently forced admin color scheme beneath the color scheme picker for admins via a notice similar to an admin notice
 * Change: Ensure admin color scheme name is lowercased before validation, saving, and comparisons
 * Change: Rename `get_forced_admin_color()` to `get_forced_color_scheme()`
+* Change: Rename `set_forced_admin_color()` to `set_forced_color_scheme()`
 * Change: Tweak attribute spacing for `label` tag
 * Change: Use HTTPS for link to WP SVN repository in bin script for configuring unit tests
 * Change: Note compatibility through WP 5.4+
 * Change: Update links to coffee2code.com to be HTTPS
 * Unit tests:
-    * New: Add tests for `add_checkbox()`, `force_admin_color()`, `hide_admin_color_input()`, `output_css()`, `register_css()`, `save_setting()`, `set_forced_admin_color()`
+    * New: Add tests for `add_checkbox()`, `force_admin_color()`, `hide_admin_color_input()`, `output_css()`, `register_css()`, `save_setting()`, `set_forced_color_scheme()`
     * Change: Remove commented out code
 * New: Add screenshots for messages indicating use of constant or filter